+++ packages/app/components/PageDialog/PageDialog.jsx
@@ -1,0 +1,60 @@
+import { Modal } from '@appsemble/react-components';
+import classNames from 'classnames';
+import PropTypes from 'prop-types';
+import React from 'react';
+
+import Block from '../Block';
+import styles from './PageDialog.css';
+
+/**
+ * Render an app page definition.
+ */
+export default class PageDialog extends React.Component {
+  static propTypes = {
+    dialog: PropTypes.shape(),
+    getBlockDefs: PropTypes.func.isRequired,
+  };
+
+  static defaultProps = {
+    dialog: null,
+  };
+
+  componentDidMount() {
+    const { dialog, getBlockDefs } = this.props;
+
+    if (dialog) {
+      getBlockDefs(dialog.blocks);
+    }
+  }
+
+  componentDidUpdate() {
+    const { dialog, getBlockDefs } = this.props;
+
+    if (dialog) {
+      getBlockDefs(dialog.blocks);
+    }
+  }
+
+  render() {
+    const { dialog, ...props } = this.props;
+
+    return (
+      <Modal isActive={!!dialog} onClose={dialog && dialog.close}>
+        {dialog && (
+          <div className={classNames('card', { [styles.fullScreen]: dialog.fullScreen })}>
+            {dialog.blocks.map((block, index) => (
+              <Block
+                // eslint-disable-next-line react/no-array-index-key
+                key={index}
+                actionCreators={dialog.actionCreators}
+                block={block}
+                data={dialog.data}
+                {...props}
+              />
+            ))}
+          </div>
+        )}
+      </Modal>
+    );
+  }
+}